refactor: パース処理を機能別にファイル分割
