+++ src/format/utils.ts
@@ -1,5 +1,5 @@
 import { relative } from "node:path";
-import type { PackageInfo } from "../parse.js";
+import type { PackageInfo } from "../parse/parseFile.js";
 
 export function createImportStatement(
 	outputPath: string,
